+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2019.02.27
+### Changed
+- Refactor edie2mjml generation
+- Add format definition with allowed attributes per block 
+
 ## [0.0.15] - 2019.02.26
 ### Changed
 - Better handle top level elements